What is Grid Terminal
This was a Blitz3D Project way back in 2007/08.  It was possible back then as another project came to fruition whcih was the Fast libraries by MixailV.

Once I started using those libraries combined with Blitz3D, I was able to do a lightweight tile & terminal system with up to 4 screens and not take too much cpu/gpu, which was incredible at the time for something done on a basic language running on DirectX 7.